feature: Make sure to stop the hover state when you mouse-leave the relation
To change the hover state of the relation, we delete the arrow SVG element on mouse enter event and recreate it.
For SVG elements that are recreated with mouse enter events, the mouse leave event will not fire depending on the timing.
Stop re-creating the SVG element and change the attributes of the SVG element instead.