fix: When I adjust the position of the end of the selected string within the span and click the expand span button, the span is not expanded.
On touch devices, the last adjusted end of the selection string becomes the focus node.
When you try to extend a span, the end in the span will not necessarily be the anchor node.
In order to support touch operations, we need to take into account that the anchor node and the focus node are swapped.