feature: When there is a difference in the number of attributes of adjacent Entities in the same row, the relationship will not be bent significantly
The position of the top edge of adjacent Entities in the same row can vary greatly depending on the number of attributes. We do not want to bend the relation too much when there is a difference in the number of attributes. We will change the condition to determine when to bend the relation significantly from the difference in the y-coordinates of the end points of the relation to the difference in the y-coordinates of the low points of the Entity.