feature: Don't reset the row height when changing modes
When the mode changes, toggle whether TypeGap is displayed depending on whether it is in simple mode or not. In this case, even if there is no change in the TypeGap value, the line height is recalculated. In order to simplify the processing on the mode side, we decided to notify TypeGap only when there is a change in the value.