feature: Stop showing icons for mobile devices on PCs with touch devices such as Surface
Essentially, we want to determine if a user can enter a shortcut key.
There is no simple way to determine this. There is a virtual keyboard on mobile devices. Whether a shortcut key can be entered on a mobile device depends on whether the virtual keyboard is displayed. Detecting this requires a hacky method such as resize events or scroll detection.
In this case, we simply used the width of the screen to distinguish between Surface and iPad.