feature: Change the keystrokes for copy and paste to match the standard OS operation
BREAKING CHANGE:
We will use the system clipboard to realize copy and paste between multiple editors. In order to access the system clipboard while keeping browser security in mind, we will handle copy and paste events for standard OS operations. In Chrome and Safafi, copy and paste events for elements other than EditableContent are fired for the Body element, so we use EditorContainer to handle events instead of individual editors.