bugfix: An error occurs when the length of the span is changed. The error occurs when there is an entity E1 with more than one relation, and there is a relation between the entity E1 and the entities on the span.
When changing the length of a span, the span is erased and rendered again. The entities on the span are deleted at the same time. When the span is erased, the span erase event fires and the position calculations for all annotations are performed. This includes the position calculation of the relation. To calculate the starting position of a relation, we need to know how many relations extend in the same direction from the entity. The position information of the entity to which the relation is connected is then referenced. However, when the length of the span is changed, the entity is temporarily deleted along with the span, so the location information cannot be obtained.